Rename inputpaneliface to InputPanelIface
Rename class files to match the other files naming.
