+++ QtFreeVirtualKeyboard/InputPanelIface.hpp
